app.component moved into components directory.
