+++ src/app/app.server.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { ServerModule } from '@angular/platform-server';
 import { AppModule } from './app.module';
-import { AppComponent } from './app.component';
+import { AppComponent } from './components/app/app.component';
 
 @NgModule({
   imports: [
