+++ src/tsconfig.app-ssr.json
@@ -1,16 +1,16 @@
 {
-    "extends": "../tsconfig.json",
-    "compilerOptions": {
-      "outDir": "../out-tsc/app",
-      "baseUrl": "./",
-      "module": "commonjs",
-      "types": []
-    },
-    "exclude": [
-      "test.ts",
-      "**/*.spec.ts"
-    ],
-    "angularCompilerOptions": {
-      "entryModule": "app/app.server.module#AppServerModule"
-    }
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "../out-tsc/app",
+    "baseUrl": "./",
+    "module": "commonjs",
+    "types": []
+  },
+  "exclude": [
+    "test.ts",
+    "**/*.spec.ts"
+  ],
+  "angularCompilerOptions": {
+    "entryModule": "app/app.server.module#AppServerModule"
+  }
 }
